First stage of unspaghettening
