--- assets/lib_Effect.lua
@@ -1,7 +1,0 @@
-function ExiWoW.Effect:buildLibrary()
-
-	local ef = ExiWoW.LibAssets.effects
-
-
-
-end
