--- collections/_index.lua
@@ -1,18 +1,0 @@
--- Library of common reusable assets
-ExiWoW.LibAssets = {}
-local req = ExiWoW.RPText.Req;	-- RPText requirement constructor
-local ty = req.Types;			-- Local filter types
-
-function ExiWoW.LibAssets.spellKitToRP(...)
-	local kits = {...}
-	local out = {}
-	for k,kit in pairs(kits) do
-		local input = ExiWoW.LibAssets.spell_kits[kit];
-		for i,v in pairs(input) do
-			out[i] = true
-		end
-	end
-	return out;
-end
-
-
