+++ db/actions.lua
@@ -1,12 +1,23 @@
 local appName, internal = ...
+local require = internal.require;
+
 -- Library for Actions --
-function ExiWoW.Action:buildLibrary()
+function internal.build.actions()
+
+	local Action = require("Action");
+	local Character = require("Character");
+	local Tools = require("Tools");
+	local UI = require("UI");
+	local RPText = require("RPText");
 
 	local ef = ExiWoW.LibAssets.effects
+	local extension = ExiWoW.R;
+	
+
 			-- LIBRARY --
 
 	-- Meta action that checks if target has ExiWoW --
-	table.insert(ExiWoW.R.actions, ExiWoW.Action:new({
+	extension:addAction({
 		id = "A",
 		global_cooldown = false,
 		suppress_all_errors = true,
@@ -25,12 +36,12 @@ function ExiWoW.Action:buildLibrary()
 				if not success then return end
 				sender = Ambiguate(sender, "all")
 				if success and UnitIsUnit(sender, "target") then
-					ExiWoW.TARGET = ExiWoW.Character:new(data, sender);
+					ExiWoW.TARGET = Character:new(data, sender);
 					local offset = 0;
 					if ExiWoW.TARGET:isFemale() then offset = 0.25;
 					elseif not ExiWoW.TARGET:isMale() then offset = 0.5; end
-					ExiWoW.Frames.targetHasExiWoWFrame.genderTexture:SetTexCoord(offset,offset+0.25,0,1);
-					ExiWoW.Frames.targetHasExiWoWFrame:Show();
+					UI.portrait.targetHasExiWoWFrame.genderTexture:SetTexCoord(offset,offset+0.25,0,1);
+					UI.portrait.targetHasExiWoWFrame:Show();
 				end
 			end
 		end,
@@ -39,10 +50,10 @@ function ExiWoW.Action:buildLibrary()
 			return true, ExiWoW.ME:export(true)
 		end
 
-	}))
+	})
 
 	-- Disrobe --
-	table.insert(ExiWoW.R.actions, ExiWoW.Action:new({
+	extension:addAction({
 		id = "DISROBE",
 		name = "Disrobe",
 		description = "Removes a piece of armor from your target.",
@@ -55,7 +66,7 @@ function ExiWoW.Action:buildLibrary()
 		cast_time = 2,
 		allow_caster_moving = false,
 		cast_sound_loop = 6425,				-- Tailoring, see http://www.wowhead.com/sound=6425/tailoring
-		max_distance = ExiWoW.Action.MELEE_RANGE,
+		max_distance = Action.MELEE_RANGE,
 		-- allow_self = false,
 		fn_cast = function(self, sender, target, suppressErrors)
 			DoEmote("KNEEL", target);
@@ -66,14 +77,14 @@ function ExiWoW.Action:buildLibrary()
 			-- Return no data, but one callback
 			return nil, function(se, success, data)
 				if not success then
-					if data and data[1] then ExiWoW:reportError(data[1], suppressErrors); end
+					if data and data[1] then Tools.reportError(data[1], suppressErrors); end
 					self:resetCooldown();
 				else
 					PlaySound(1202, "SFX");
-					ExiWoW:reportError(
-						ExiWoW:unitRpName(sender) .. " successfully removed "..
+					Tools.reportError(
+						Tools.unitRpName(sender) .. " successfully removed "..
 						Ambiguate(UnitName(target), "all").."'s "..
-						ExiWoW:itemSlotToname(data.slot).."!"
+						Tools.itemSlotToname(data.slot).."!"
 					);
 				end
 			end
@@ -106,18 +117,18 @@ function ExiWoW.Action:buildLibrary()
 			end
 
 			local slot = equipped_slots[ math.random( #equipped_slots ) ];
-			ExiWoW.Character:removeEquipped(slot);
-			ExiWoW:reportError(ExiWoW:unitRpName(sender) .. " tugged off your "..ExiWoW:itemSlotToname(slot).."!");
+			Character:removeEquipped(slot);
+			Tools.reportError(Tools.unitRpName(sender) .. " tugged off your "..Tools.itemSlotToname(slot).."!");
 			if not UnitIsUnit(Ambiguate(sender, "ALL"), "player") then 
 				PlaySound(1202, "SFX");
 			end
 			return true, {slot=slot}
 		end
 
-	}))
+	})
 
 	-- meditate --
-	table.insert(ExiWoW.R.actions, ExiWoW.Action:new({
+	extension:addAction({
 		id = "MEDITATE",
 		self_only = true,
 		name = "Meditate",
@@ -131,69 +142,69 @@ function ExiWoW.Action:buildLibrary()
 		fn_receive = function(self, sender, target, suppressErrors)
 
 			if ExiWoW.ME.meditating then
-				return ExiWoW:reportError("You are already meditating!");
+				return Tools.reportError("You are already meditating!");
 			end
 
 			-- Start meditation --
 			DoEmote("SIT");
 			ExiWoW.ME.meditating = true;
 			ExiWoW.ME:toggleResting(true)
-			ExiWoW.Character:bind("PLAYER_STARTED_MOVING", function()
+			Character.bind("PLAYER_STARTED_MOVING", function()
 				ExiWoW.ME:toggleResting(false)
 				ExiWoW.ME.meditating = false;
 			end, 1);
 			return true
 
 		end
-	}))
+	})
 
 	-- Spot excitement (Public, melee range) --
-	table.insert(ExiWoW.R.actions, ExiWoW.Action:new({
+	extension:addAction({
 		id = "ASSESS",
 		name = "Assess",
 		important = true,
 		description = "Take a good look at your target, revealing some information about them.",
 		texture = "inv_darkmoon_eye",
 		cooldown = 0,
-		max_distance = ExiWoW.Action.MELEE_RANGE,
+		max_distance = Action.MELEE_RANGE,
 		party_restricted = false,
 		fn_send = function(self, sender, target, suppressErrors)
 			-- We only need a callback for this
-			return nil, function(se, success, data) ExiWoW.Action:handleInspectCallback(target, success, data) end
+			return nil, function(se, success, data) Action.handleInspectCallback(target, success, data) end
 		end,
 		fn_receive = function()
 			return true, ExiWoW.ME:export(true)
 		end
-	}));
+	});
 
 	-- Sniff (Worgen) --
-	table.insert(ExiWoW.R.actions, ExiWoW.Action:new({
+	extension:addAction({
 		id = "SNIFF",
 		name = "Sniff",
 		description = "Sniff out some information about your target from a distance.",
 		texture = "inv_wolfdraenormountshadow",
 		cooldown = 0,
-		max_distance = ExiWoW.Action.CASTER_RANGE,
+		max_distance = Action.CASTER_RANGE,
 		party_restricted = false,
 		allowed_races = {"Worgen"},
 		fn_send = function(self, sender, target, suppressErrors)
 			DoEmote("SNIFF", target);
 			-- Callback
-			return nil, function(se, success, data) ExiWoW.Action:handleInspectCallback(target, success, data) end
+			return nil, function(se, success, data) Action.handleInspectCallback(target, success, data) end
 		end,
 		fn_receive = function()
 			return true, ExiWoW.ME:export(true)
 		end
-	}));
+	});
 
 	-- Tickle --
-	table.insert(ExiWoW.R.actions, ExiWoW.Action:new({
+	extension:addAction({
 		id = "TICKLE",
 		name = "Tickle",
 		description = "Tickle a player.",
 		texture = "Spell_shadow_fingerofdeath",
 		cooldown = 6,
-		max_distance = ExiWoW.Action.MELEE_RANGE,
+		max_distance = Action.MELEE_RANGE,
 		party_restricted = false,
 		fn_send = function(self, sender, target, suppressErrors)
 			if not UnitIsUnit(target, "player") then 
@@ -206,16 +217,16 @@ function ExiWoW.Action:buildLibrary()
 			self:receiveRPText(sender, target, args)
 			return true
 		end
-	}));
+	});
 
 	-- Wedgie --
-	table.insert(ExiWoW.R.actions, ExiWoW.Action:new({
+	extension:addAction({
 		id = "WEDGIE",
 		name = "Wedgie",
 		description = "Give a player a wedgie, provided they're wearing underwear.",
 		texture = "Spell_holy_fistofjustice",
 		cooldown = 6,
-		max_distance = ExiWoW.Action.MELEE_RANGE,
+		max_distance = Action.MELEE_RANGE,
 		target_has_underwear = true,
 		party_restricted = false,
 		fn_send = function(self, sender, target, suppressErrors)
@@ -233,10 +244,10 @@ function ExiWoW.Action:buildLibrary()
 			self:receiveRPText(sender, target, args);
 			return true
 		end
-	}));
+	});
 
 	-- Forage --
-	table.insert(ExiWoW.R.actions, ExiWoW.Action:new({
+	extension:addAction({
 		id = "FORAGE",
 		name = "Forage",
 		description = "Search your active area for items.",
@@ -253,13 +264,13 @@ function ExiWoW.Action:buildLibrary()
 			return nil;
 		end,
 		fn_receive = function(self, sender, target, args)
-			ExiWoW.Character:forage()
+			Character.forage()
 			return true
 		end
-	}));
+	});
 
 	-- Pace --
-	table.insert(ExiWoW.R.actions, ExiWoW.Action:new({
+	extension:addAction({
 		id = "PACE",
 		name = "Pace",
 		description = "First use: Stake a starting point. Second use: Get the distance from the starting point, measured in map coordinates.",
@@ -280,16 +291,16 @@ function ExiWoW.Action:buildLibrary()
 				local y = self.starting_point.y
 				local dist = math.floor(math.sqrt((px-x)*(px-x)+(py-y)*(py-y))*100)/100
 				self.starting_point = nil
-				ExiWoW.RPText:print("You are comfortable that you paced a distance of "..dist.." units")
+				RPText.print("You are comfortable that you paced a distance of "..dist.." units")
 				PlaySound(73276, "SFX")
 			else
 				self.starting_point = {x=px, y=py}
 				PlaySound(42485, "SFX")
-				ExiWoW.RPText:print("You stake a starting point at X:"..(math.floor(px*100)/100)..", Y:"..(math.floor(py*100)/100))
+				RPText.print("You stake a starting point at X:"..(math.floor(px*100)/100)..", Y:"..(math.floor(py*100)/100))
 			end
 			return true
 		end
-	}));
+	});
 
 
 
@@ -301,7 +312,7 @@ function ExiWoW.Action:buildLibrary()
 
 
 		-- Consumable --
-	table.insert(ExiWoW.R.actions, ExiWoW.Action:new({
+	extension:addAction({
 		id = "THROW_SAND",
 		name = "Sand",
 		description = "Throw sand at your target.",
@@ -324,10 +335,10 @@ function ExiWoW.Action:buildLibrary()
 			ef.addExcitementMasochisticDefault();
 			return true
 		end
-	}));
+	});
 
 	-- Claw pinch
-	table.insert(ExiWoW.R.actions, ExiWoW.Action:new({
+	extension:addAction({
 		id = "CLAW_PINCH",
 		name = "Claw Pinch",
 		description = "Use your large claw to pinch your target.",
@@ -348,7 +359,7 @@ function ExiWoW.Action:buildLibrary()
 			ef.addExcitementMasochisticDefault();
 			return true
 		end
-	}));
+	});
 	
 
 end
