+++ db/effects.lua
@@ -1,0 +1,9 @@
+local appName, internal = ...
+local require = internal.require;
+
+function internal.build.effects()
+
+	--
+
+
+end
