+++ db/rptexts.lua
@@ -1,30 +1,29 @@
-function ExiWoW.RPText:buildLibrary()
+local appName, internal = ...
+local require = internal.require;
+
+function internal.build.rptexts()
 	
-	local req = ExiWoW.RPText.Req;	-- RPText requirement constructor
-	local ty = req.Types;			-- Local filter types
+	local Condition = require("Condition");	-- RPText requirement constructor
+	local ty = Condition.Types;			-- Local filter types
 	local assetLib = ExiWoW.LibAssets;
 	local spellKits = assetLib.spell_kits;		-- RP Text spell kits library
 	local rtc = ExiWoW.LibAssets.rpTextConds;	-- RP Text conditions
 	local ef = ExiWoW.LibAssets.effects;		-- RP Text effects library
-	
+	local RPText = require("RPText");
+
 	-- Root extension
 	local R = ExiWoW.R.rpTexts;
-	
-
-	-- Gets a formatted spell kit from lib_Assets (or more)
-	-- A spell kit is a collection of spell names that share the same theme, such as frost, fire, basilisk stun etc
-	local getsk = ExiWoW.LibAssets.spellKitToRP;
 		
 	-- Only wholesome family friendly PG stuff in here
 
 	-- Tickle target
-	table.insert(R, ExiWoW.RPText:new({
+	table.insert(R, RPText:new({
 		id = "TICKLE",
 		text_sender = "You tickle %T!",
 		text_receiver = "%T tickles you!",
 		requirements = {},
 	}))
-	table.insert(R, ExiWoW.RPText:new({
+	table.insert(R, RPText:new({
 		id = "TICKLE",
 		text_sender = "You tickle %T with your %item!",
 		text_receiver = "%S tickles you with %Shis %item!",
@@ -32,14 +31,14 @@ function ExiWoW.RPText:buildLibrary()
 	}))
 
 	-- Tickle self
-	table.insert(R, ExiWoW.RPText:new({
+	table.insert(R, RPText:new({
 		id = "TICKLE",
 		text_bystander = "%T tickles %Thimself!",
 		text_receiver = "You tickle yourself!",
 		requirements = {},
 		--fn = ef.addExcitementMasochisticDefault
 	}))
-	table.insert(R, ExiWoW.RPText:new({
+	table.insert(R, RPText:new({
 		id = "TICKLE",
 		text_bystander = "%T tickles %Thimself with %This %item!",
 		text_receiver = "You use your %item to tickle yourself!",
@@ -48,7 +47,7 @@ function ExiWoW.RPText:buildLibrary()
 	}))
 
 	-- Wedgie
-	table.insert(R, ExiWoW.RPText:new({
+	table.insert(R, RPText:new({
 		id = "WEDGIE",
 		text_bystander = "%S grabs a hold of %T's %Tundies, giving %Thim a wedgie!",
 		text_sender = "You grab a hold of %T's %Tundies, giving %Thim a wedgie!",
@@ -58,7 +57,7 @@ function ExiWoW.RPText:buildLibrary()
 		--fn = ef.addExcitementMasochisticDefault
 	}))
 
-	table.insert(R, ExiWoW.RPText:new({
+	table.insert(R, RPText:new({
 		id = "WEDGIE",
 		text_bystander = "%T grabs a hold of %This %Tundies and gives %Thimself a wedgie!",
 		text_receiver = "You grab a hold of your %Tundies and give yourself a wedgie!",
@@ -69,8 +68,8 @@ function ExiWoW.RPText:buildLibrary()
 	
 
 	-- Insect swarm
-	table.insert(R, ExiWoW.RPText:new({
-		id = getsk("insects"),
+	table.insert(R, RPText:new({
+		id = spellKits.insects,
 		text_receiver = "The insects get into your equipment, skittering across your body!",
 		--sound = 35103,
 		requirements = {rtc.spellAdd},
@@ -78,23 +77,23 @@ function ExiWoW.RPText:buildLibrary()
 	}))
 
 	-- Sand ADD
-	table.insert(R, ExiWoW.RPText:new({
-		id = getsk("sand"),
+	table.insert(R, RPText:new({
+		id = spellKits.sand,
 		text_receiver = "Sand gets into your clothes!",
 		--sound = 35103,
 		requirements = {rtc.spellAdd},
 		fn = ef.addExcitementMasochisticDefault
 	}))
 
-	table.insert(R, ExiWoW.RPText:new({
-		id = getsk("dirt"),
+	table.insert(R, RPText:new({
+		id = spellKits.dirt,
 		text_receiver = "Some dirt gets into your clothes!",
 		--sound = 35103,
 		requirements = {rtc.spellTick},
 		fn = ef.addExcitementMasochisticDefault
 	}))
 
-	table.insert(R, ExiWoW.RPText:new({
+	table.insert(R, RPText:new({
 		id = "THROW_SAND",
 		text_bystander = "%T throws sand into the air, some of which falls back down on %Thim and into %This clothes!",
 		text_receiver = "You throw sand into the air, some of which falls back down on you and into your clothes!",
@@ -102,7 +101,7 @@ function ExiWoW.RPText:buildLibrary()
 		requirements = {}
 	}))
 
-	table.insert(R, ExiWoW.RPText:new({
+	table.insert(R, RPText:new({
 		id = "THROW_SAND",
 		text_bystander = "%S throws a handful of sand at %T!",
 		text_sender = "You throw a handful of sand at %T!",
@@ -111,7 +110,7 @@ function ExiWoW.RPText:buildLibrary()
 		requirements = {}
 	}))
 
-	table.insert(R, ExiWoW.RPText:new({
+	table.insert(R, RPText:new({
 		id = "CLAW_PINCH",
 		text_bystander = "%S pinches %T's side with a big claw!",
 		text_sender = "You pinch %T's side with your big claw!",
@@ -119,7 +118,7 @@ function ExiWoW.RPText:buildLibrary()
 		sound = 36721,
 		requirements = {}
 	}))
-	table.insert(R, ExiWoW.RPText:new({
+	table.insert(R, RPText:new({
 		id = "CLAW_PINCH",
 		text_bystander = "%T pinches %T's nose with a big claw!",
 		text_sender = "You pinch %T's nose with your big claw!",
@@ -127,7 +126,7 @@ function ExiWoW.RPText:buildLibrary()
 		sound = 36721,
 		requirements = {}
 	}))
-	table.insert(R, ExiWoW.RPText:new({
+	table.insert(R, RPText:new({
 		id = "CLAW_PINCH",
 		text_bystander = "%T pinches %This own nose with a big claw!",
 		text_receiver = "You pinch your nose with your big claw!",
