+++ lib/npc.lua
