+++ lib/spells.lua
